Data objects refactor
