+++ darwin/future/data_objects/team_member_role.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class TeamMemberRole(Enum):
+    TEAM_OWNER = "owner"
+    TEAM_ADMIN = "admin"
+    USER = "member"
+    WORKFORCE_MANAGER = "workforce_manager"
+    WORKER = "annotator"
